Initial q space plotting
