+++ Project.toml
@@ -4,6 +4,7 @@ authors = ["The Sunny team"]
 version = "0.7.4"
 
 [deps]
+Brillouin = "23470ee3-d0df-4052-8b1a-8cbd6363e7f0"
 CrystalInfoFramework = "6007d9b0-c6b2-11e8-0510-1d10e825f3f1"
 DynamicPolynomials = "7c1d4256-1411-5781-91ec-d7bc3513ac07"
 ElasticArrays = "fdbdab4c-e67f-52f5-8c3f-e7b388dad3d4"
@@ -35,6 +36,7 @@ PlottingExt = "Makie"
 WGLMakiePrecompilesExt = "WGLMakie"
 
 [compat]
+Brillouin = "0.5.19"
 CrystalInfoFramework = "0.7.0"
 DynamicPolynomials = "0.6.0"
 ElasticArrays = "1.2.12"
